% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{filter2}
\alias{filter2}
\title{Title}
\usage{
filter2(df, condition)
}
\arguments{
\item{df}{A Data frame}

\item{i}{a vector}
}
\value{
A subset data.frame with less rows
}
\description{
Title
}
\examples{
filter2(iris,3:5)
}
