% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minidplyr4-package.R
\docType{package}
\name{minidplyr4-package}
\alias{minidplyr4}
\alias{minidplyr4-package}
\title{minidplyr4: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/privefl/minidplyr4}
  \item Report bugs at \url{https://github.com/privefl/minidplyr4/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
