% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Subset columns of a data frame}
\usage{
select(x, var)
}
\arguments{
\item{x}{A data frame}

\item{var}{Either a vector of column indices or variable names or logicals}
}
\value{
A subset data frame
}
\description{
Subset columns of a data frame
}
\examples{
select(iris, 1:2)
select(iris, "Species")
select(iris, c(TRUE, FALSE, FALSE, TRUE, FALSE))
}
