% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select2.R
\name{select2}
\alias{select2}
\title{Title
Select variables}
\usage{
select2(df, var)
}
\arguments{
\item{df}{A data frame}

\item{var}{A column in the data frame or a vector}
}
\value{
A subsetted data frame
}
\description{
Title
Select variables
}
\examples{
select2(iris, 1)
select2(iris, 1:3)
}
