% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2.R
\name{filter2}
\alias{filter2}
\title{Filtering columns from a dataframe}
\usage{
filter2(x, var)
}
\arguments{
\item{x}{Dataframe}

\item{var}{Number of clumns you want to filter off}
}
\value{
A subsetted dataframe
}
\description{
Filtering columns from a dataframe
}
\examples{
filter2(iris, 1:2)
filter2(iris, 3:5)
}
