% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2.R
\name{filter2}
\alias{filter2}
\title{Title}
\usage{
filter2(x)
}
\arguments{
\item{x}{A vector}

\item{...}{}
}
\value{

}
\description{
The description
}
\details{
\strong{These} \emph{are} the \code{details}
\itemize{
\item list element 1
\item list element 2
}
}
\examples{

}
\seealso{
\link{Google}(google.com][test()]
}
