% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter2}
\alias{filter2}
\title{Filter rows of a data frame.}
\usage{
filter2(df, ind)
}
\arguments{
\item{df}{A data frame.}

\item{ind}{A vector of indices.}
}
\value{
A subsetted data frame with less rows.
}
\description{
Filter rows of a data frame.
}
\examples{
filter2(iris, 3:5)
}
