% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select a column in a dataframe}
\usage{
select(x, var)
}
\arguments{
\item{x}{Dataframe}

\item{var}{Column number of the dataframe that you want to extract}
}
\value{
A subsetted dataframe
}
\description{
Select a column in a dataframe
}
\examples{
iris
}
