% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select function}
\usage{
select(df, i)
}
\arguments{
\item{df}{A data frame}

\item{i}{A column}
}
\value{
Sumbitted data.frame column
}
\description{
Select function
}
\examples{
select(iris,"name")
}
