% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_2.R
\name{select2}
\alias{select2}
\title{Select 2}
\usage{
select2(df, names_or_ind)
}
\arguments{
\item{df}{a dataframe}

\item{names_or_ind}{A vector of column names or of column indicies.}
}
\value{
Antoher data frame, with the subsetted columns.
}
\description{
this function is like select 2, but just a bit different
}
\details{
\strong{details:}
\itemize{
\item element 1
\item element 2
}
}
\examples{
res <. select2(iris, 2)
plot(res)
}
