% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select2}
\alias{select2}
\title{Select variables}
\usage{
select2(df, ind)
}
\arguments{
\item{df}{A data frame.}

\item{ind}{Either a vector of variable indices, names, or a boolean vector.}
}
\value{
A subsetted data frame.
}
\description{
Select a subset of variables from a data frame.
}
\examples{
select2(iris, 1)
select2(iris, 1:3)
select2(iris, -1)
select2(iris, c("Sepal.Width", "Species"))
select2(iris, c(TRUE, FALSE, FALSE, TRUE, FALSE))
}
